DBZ-8134 Fix parsing of server ID

+++ debezium-connector-binlog/src/main/java/io/debezium/connector/binlog/BinlogConnectorConfig.java
@@ -832,7 +832,7 @@ public abstract class BinlogConnectorConfig extends HistorizedRelationalDatabase
      * @return the database cluster server unique identifier
      */
     public long getServerId() {
-        return config.getInteger(SERVER_ID);
+        return config.getLong(SERVER_ID);
     }
 
     /**
